# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: LESS API
  version: 0.1.0
  description: |
    LESSのREST API定義。本fileがフロントエンドとバックエンドの契約の正本である。

    現在のスコープは以下とする。
      - Phase 0: 認証 (F-001)
      - Phase 1: Category初期data / Item CRUD / 一覧filter / Tag / 使用記録 (F-003〜F-006)
      - Phase 2: StorageUnit / 収納割当 / 容量集計 / 携行区分 (F-008 / F-009)

    Review (Phase 3)、PurchaseCandidate (Phase 4)、Scenario (Phase 5)、
    Import/Export (Phase 6) はまだ定義しない。

    Phase 2で設計書と異なる判断をした点:
      - 更新は 設計書 12.4 の `PUT /api/storage-units/{publicId}` に従い全置換とする。
        PATCHによる部分更新は導入しない (Itemの `updateItem` と同じ方針)。
      - 重量・容積のfield名は設計書 13.7 / 13.8 のcolumn名に合わせ単数形
        (`tareWeightGram` / `maximumVolumeMilliliter` 等) とする。
        Phase 1の `weightGram` / `volumeMilliliter` と揃える。
      - 設計書 12.4 に無い `restore` / `contents` / 割当の個別操作 /
        `GET /items/{publicId}/storage-allocations` を追加した。
        いずれもPhase 2の要件 (収納内容編集・分割割当・未割当数量) に必要である。
      - 収納単位の親集計は設計書 16.2 が子孫を規定していないため、
        直接割当分 (`itemWeightGram`) と子孫分 (`descendantWeightGram`) を
        別fieldで返し、`totalWeightGram` で合算する。二重計上を避けるため
        子孫分には子孫自身の自重を含め、親の自重は含めない。

    共通仕様 (設計書 12.1):
      - base path: `/api`
      - Content-Type: `application/json`
      - datetime: RFC3339 (UTC)
      - JSON field: camelCase
      - enum value: lowercase snake_case
      - path: 複数形kebab-case
      - URLでは内部IDではなくpublicIdを使用する

    一覧のpagination:
      設計書 12.2 はcursor paginationを基本とするが、Item一覧は
      任意のsort keyとの組み合わせ、および総件数表示を要件とするため
...

...

...
servers:
...
paths:
...
  /items:
    get:
      tags: [items]
      operationId: listItems
      summary: 所持品一覧を取得する
      description: |
        認証ユーザーの所持品を検索・絞り込み・並び替えして返す (設計書 9.4)。
        archive済みのアイテムは既定で除外し、`includeDeleted=true` の場合のみ含める。
      security:
        - sessionCookie: []
      parameters:
        - name: keyword
          in: query
          description: アイテム名またはメモの部分一致 (大文字小文字を区別しない)
          schema:
            type: string
            maxLength: 100
        - name: categoryPublicId
          in: query
          description: カテゴリーで絞り込む
          schema:
            type: string
            format: uuid
        - name: tagPublicId
          in: query
          description: タグで絞り込む
          schema:
            type: string
            format: uuid
        - name: necessityLevelCode
          in: query
          schema:
            $ref: "#/components/schemas/NecessityLevelCode"
        - name: usageFrequencyCode
          in: query
          schema:
            $ref: "#/components/schemas/UsageFrequencyCode"
        - name: mobilityClassCode
          in: query
          schema:
            $ref: "#/components/schemas/MobilityClassCode"
        - name: storageUnitPublicId
          in: query
          description: |
            指定した収納単位へ直接割当されているアイテムだけを返す (設計書 9.4)。
            子収納単位の内容は含めない。
          schema:
            type: string
            format: uuid
        - name: isUnassigned
          in: query
          description: |
            trueの場合、未割当数量が1以上のアイテムだけを返す (設計書 9.4)。
            未割当数量 = quantity - 収納割当数量の合計。
          schema:
            type: boolean
        - name: includeDeleted
          in: query
          description: archive済みのアイテムを含めるか。未指定時はfalse。
          schema:
            type: boolean
        - name: sort
          in: query
          description: 並び替えkey。未指定時は updatedAt。
          schema:
            $ref: "#/components/schemas/ItemSortKey"
        - name: order
          in: query
          description: 並び順。未指定時は desc。
          schema:
            $ref: "#/components/schemas/SortOrder"
        - $ref: "#/components/parameters/LimitQueryParameter"
        - $ref: "#/components/parameters/OffsetQueryParameter"
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [items]
      operationId: createItem
      summary: 所持品を登録する
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: 登録に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: |
            指定したカテゴリーまたはタグが存在しない。
            他ユーザーのカテゴリー・タグを指定した場合も本statusを返す (設計書 18.3)。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{publicId}:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [items]
      operationId: getItemByPublicId
      summary: 所持品を取得する
      description: |
        他ユーザーのpublicIdを指定した場合も 404 を返し、
        存在有無を公開しない (設計書 18.3)。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [items]
      operationId: updateItem
      summary: 所持品を更新する
      description: |
        全項目を置き換える。省略した任意項目はNULLへ更新される。
        `expectedVersion` が現在versionと一致しない場合は 409 を返す。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: 更新に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{publicId}/archive:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    post:
      tags: [items]
      operationId: archiveItem
      summary: 所持品をarchiveする
      description: |
        soft deleteとして `archivedAt` を設定する (設計書 1.4)。
        archive済みのアイテムは既定の一覧から除外され、使用記録を追加できない。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemVersionRequest"
      responses:
        "200":
          description: archiveに成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{publicId}/restore:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    post:
      tags: [items]
      operationId: restoreItem
      summary: archive済みの所持品を復元する
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemVersionRequest"
      responses:
        "200":
          description: 復元に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{publicId}/usage-records:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [items]
      operationId: listItemUsageRecords
      summary: 使用記録の履歴を取得する
      description: 使用日時の降順で返す。
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/LimitQueryParameter"
        - $ref: "#/components/parameters/OffsetQueryParameter"
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemUsageRecordListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [items]
      operationId: createItemUsageRecord
      summary: 使用記録を登録する
      description: |
        登録と同時にアイテムの最終使用日時を更新する。
        既存の最終使用日時より古い記録では更新しない。
        archive済みのアイテムへは登録できず 422 を返す (設計書 7.2)。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemUsageRecordRequest"
      responses:
        "201":
          description: 登録に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemUsageRecordResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: less_session
      description: |
        session token Cookie。HttpOnly、SameSite=Lax、本番環境ではSecureを付与する。
        実際のCookie名は環境変数 `SESSION_COOKIE_NAME` で変更できる。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token。非HttpOnly Cookie `less_csrf` の値をそのまま送信する
        (signed double-submit cookie)。state変更requestで必須とする。
  parameters:
...
    LimitQueryParameter:
      name: limit
      in: query
      description: 取得件数。未指定時は50。
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
    OffsetQueryParameter:
      name: offset
      in: query
      description: 読み飛ばす件数。未指定時は0。
      schema:
        type: integer
        format: int32
        minimum: 0
  responses:
    BadRequest:
      description: request形式が不正である
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: BAD_REQUEST
            message: リクエストの形式が正しくありません。
            fieldErrors:
              - field: email
                code: INVALID_FORMAT
                message: メールアドレスの形式が正しくありません。
            requestId: req_01JQ0000000000000000000000
    Unauthorized:
      description: 未認証である
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHENTICATED
            message: ログインが必要です。
            fieldErrors: []
            requestId: req_01JQ0000000000000000000000
    Forbidden:
      description: CSRF検証に失敗した、または操作が許可されていない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: CSRF_TOKEN_INVALID
            message: セキュリティトークンが無効です。画面を再読み込みしてください。
            fieldErrors: []
            requestId: req_01JQ0000000000000000000000
...
    InternalServerError:
      description: 予期しないerrorが発生した
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INTERNAL_SERVER_ERROR
            message: サーバーでエラーが発生しました。
            fieldErrors: []
            requestId: req_01JQ0000000000000000000000
  schemas:
...
    NecessityLevelCode:
      type: string
      description: 必要度 (設計書 14.5)
      enum: [essential, important, optional, undecided, unnecessary]
      example: essential
    UsageFrequencyCode:
      type: string
      description: 使用頻度 (設計書 14.3)
      enum: [daily, weekly, monthly, quarterly, yearly, rarely, never]
      example: monthly
...
    MobilityClassCode:
      type: string
      description: 携行区分 (設計書 16.1)
      enum:
        [
          worn,
          pocket,
          daily_bag,
          on_demand,
          self_carry,
          parcel,
          mover,
          dispose_rebuy,
          fixed,
        ]
      example: daily_bag
...
    ItemSortKey:
      type: string
      description: 所持品一覧の並び替えkey
      enum: [name, quantity, lastUsedAt, updatedAt]
      example: updatedAt
...
    SortOrder:
      type: string
      enum: [asc, desc]
      example: desc

    # ---------------------------------------------------------------------
    # 共通
    # ---------------------------------------------------------------------
...
    ItemResponse:
      type: object
      description: |
        所持品 (設計書 12.6 / 13.7)。内部IDを含めない。
        一覧と詳細で同一の表現を使用する。

        `review` (見直しスコア) はPhase 3のスコープのため含めない。
      required:
        - publicId
        - name
        - category
        - itemKindCode
        - itemKindLabel
        - quantity
        - desiredQuantity
        - unitName
        - necessityLevelCode
        - necessityLevelLabel
        - usageFrequencyCode
        - usageFrequencyLabel
        - substitutabilityCode
        - substitutabilityLabel
        - mobilityClassCode
        - mobilityClassLabel
        - ownershipReason
        - disposalCondition
        - lastUsedAt
        - purchasedOn
        - purchaseAmount
        - replacementAmount
        - resaleAmount
        - weightGram
        - volumeMilliliter
        - isFragile
        - isValuable
        - isSentimental
        - requiresMaintenance
        - expiresOn
        - sourceUrl
        - notes
        - isConfirmed
        - confirmedAt
        - tags
        - storageAllocations
        - unassignedQuantity
        - isArchived
        - archivedAt
        - version
        - createdAt
        - updatedAt
      additionalProperties: false
      properties:
        publicId:
          type: string
          format: uuid
        name:
          type: string
          example: 折りたたみ傘
        category:
          $ref: "#/components/schemas/CategoryReferenceResponse"
        itemKindCode:
          $ref: "#/components/schemas/ItemKindCode"
        itemKindLabel:
          type: string
          example: 耐久品
        quantity:
          type: integer
          format: int32
          example: 1
        desiredQuantity:
          type: integer
          format: int32
          nullable: true
          description: 希望上限数量。未設定の場合はnull。
          example: 1
        unitName:
          type: string
          example: 本
        necessityLevelCode:
          $ref: "#/components/schemas/NecessityLevelCode"
        necessityLevelLabel:
          type: string
          example: 必須
        usageFrequencyCode:
          $ref: "#/components/schemas/UsageFrequencyCode"
        usageFrequencyLabel:
          type: string
          example: 月に1回程度
        substitutabilityCode:
          $ref: "#/components/schemas/SubstitutabilityCode"
        substitutabilityLabel:
          type: string
          example: 代替不可
        mobilityClassCode:
          $ref: "#/components/schemas/MobilityClassCode"
        mobilityClassLabel:
          type: string
          example: 常時リュック
        ownershipReason:
          type: string
          nullable: true
          example: 突然の雨に対応するため
        disposalCondition:
          type: string
          nullable: true
          example: 破損して修理不能になった場合
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        purchasedOn:
          type: string
          format: date
          nullable: true
        purchaseAmount:
          type: integer
          format: int64
          nullable: true
          description: 購入金額 (円)。丸め誤差を避けるため整数で扱う (設計書 11章)。
        replacementAmount:
          type: integer
          format: int64
          nullable: true
        resaleAmount:
          type: integer
          format: int64
          nullable: true
        weightGram:
          type: integer
          format: int32
          nullable: true
        volumeMilliliter:
          type: integer
          format: int32
          nullable: true
        isFragile:
          type: boolean
        isValuable:
          type: boolean
        isSentimental:
          type: boolean
        requiresMaintenance:
          type: boolean
        expiresOn:
          type: string
          format: date
          nullable: true
        sourceUrl:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
        notes:
          type: string
          nullable: true
        isConfirmed:
          type: boolean
          description: |
            棚卸し確認済みか (設計書 13.7)。
            確認操作 (`confirmItem`) はPhase 1のスコープ外のため、常にfalseとなる。
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagReferenceResponse"
        storageAllocations:
          type: array
          description: |
            本アイテムがどの収納単位へ何個入っているか (F-009)。
            同一アイテムを複数収納単位へ分割割当できるため配列とする。
          items:
            $ref: "#/components/schemas/ItemStorageAllocationResponse"
        unassignedQuantity:
          type: integer
          format: int32
          description: |
            未割当数量。`quantity - sum(storageAllocations[].quantity)` で算出する。
            DBへ重複保存せず取得時に算出する。
          example: 0
        isArchived:
          type: boolean
          description: archive (soft delete) 済みか
        archivedAt:
          type: string
          format: date-time
          nullable: true
          description: archive日時。DBの `deleted_at` に対応する。
        version:
          type: integer
          format: int32
          description: 楽観ロック用のversion (設計書 11.7)
          example: 3
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ItemListResponse:
      type: object
      required: [items, pagination]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"
    CreateItemRequest:
      type: object
      description: 設計書 12.5 に対応する。
      required:
        - name
        - categoryPublicId
        - quantity
        - necessityLevelCode
        - usageFrequencyCode
        - substitutabilityCode
        - mobilityClassCode
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: 折りたたみ傘
        categoryPublicId:
          type: string
          format: uuid
        itemKindCode:
          allOf:
            - $ref: "#/components/schemas/ItemKindCode"
          description: 未指定時はserverが `durable` を適用する。
        quantity:
          type: integer
          format: int32
          minimum: 0
          maximum: 1000000
          example: 1
        desiredQuantity:
          type: integer
          format: int32
          minimum: 0
          maximum: 1000000
          nullable: true
          example: 1
        unitName:
          type: string
          minLength: 1
          maxLength: 20
          description: 未指定時はserverが `個` を適用する。
          example: 本
        necessityLevelCode:
          $ref: "#/components/schemas/NecessityLevelCode"
        usageFrequencyCode:
          $ref: "#/components/schemas/UsageFrequencyCode"
        substitutabilityCode:
          $ref: "#/components/schemas/SubstitutabilityCode"
        mobilityClassCode:
          $ref: "#/components/schemas/MobilityClassCode"
        ownershipReason:
          type: string
          maxLength: 1000
          nullable: true
        disposalCondition:
          type: string
          maxLength: 1000
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        purchasedOn:
          type: string
          format: date
          nullable: true
        purchaseAmount:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        replacementAmount:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        resaleAmount:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        weightGram:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        volumeMilliliter:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        isFragile:
          type: boolean
          description: 未指定時はfalse。
        isValuable:
          type: boolean
          description: 未指定時はfalse。
        isSentimental:
          type: boolean
          description: 未指定時はfalse。
        requiresMaintenance:
          type: boolean
          description: 未指定時はfalse。
        expiresOn:
          type: string
          format: date
          nullable: true
        sourceUrl:
          type: string
          maxLength: 2048
          nullable: true
          description: httpまたはhttpsのみを許可する (設計書 24.15)。
        notes:
          type: string
          maxLength: 2000
          nullable: true
        tagPublicIds:
          type: array
          maxItems: 30
          items:
            type: string
            format: uuid
...